style: moved csv saving funcs to utils
